requirements.txt
Flask==3.0.0
openai==1.59.3
python-dotenv==1.0.0
requests==2.32.3
gunicorn